+++ library/src/main/java/org/dbtools/android/commons/ui/compose/ContentAlpha.kt
@@ -1,0 +1,5 @@
+package org.dbtools.android.commons.ui.compose
+
+object ContentAlpha {
+    const val DISABLED: Float = 0.38f
+}